+++ 05-SceneManager/Portal.cpp
@@ -1,13 +1,11 @@
 #include "Portal.h"
 
 
-CPortal::CPortal(float l, float t, float r, float b, int scene_id, int nextX, int nextY ): CGameObject(x,y)
+CPortal::CPortal(float l, float t, float r, float b, int scene_id, float nextX, float nextY ): CGameObject(x,y)
 {
 	this->scene_id = scene_id;
 	x = l; 
 	y = t;
-	//width = r - l;
-	//height = b - t;
 	next_x = nextX;
 	next_y = nextY;
 }
